* 'master' of git://scm.dev.nokia.troll.no/qt/qt-fire-team: (108 commits)
  Fix horizontal centered text drawing without word wrap
  Remove Font Smoothing hack for OS X
  Revert "Fixed underline offset after raster coordinate system change."
  Fix the painting for widgets painting directly on screen.
  Fixed bad merges in GL windowsurface and get rid of redundant variable.
  Added WindowSurfaceFeature enum to QWindowSurface.
  Added preserved contents property to QWindowSurface.
  Fixed underline offset after raster coordinate system change.
  Designer: Fix tab order in action editor dialog.
  tst_qlocale: improve failure message in tst_QLocale::macDefaultLocale
  Add a once-over unrolled fromLatin1 conversion (32 characters)
  Add 16-byte loads of the Neon fromLatin1 functions
  Fix warning about unused parameter
  Make QtUiTools use the lib-infix passed to configure.
  Skip failing tests
  Removed old autotest references from Symbian build files
  Icon in QMessageBox::question is deformed on Symbian
  Prevent crash when connecting signal in splitview
  GraphicsViews: Focused widget is not visible after orientation change
  GraphicsViews: Focused widget is not visible after orientation change
  ...

+++ tests/benchmarks/corelib/io/qfile/main.cpp
@@ -60,16 +60,16 @@
 
 // 10 predefined (but random() seek positions
 // hardcoded to be comparable over several runs
-const int seekpos[] = {TF_SIZE*0.52,
-                       TF_SIZE*0.23,
-                       TF_SIZE*0.73,
-                       TF_SIZE*0.77,
-                       TF_SIZE*0.80,
-                       TF_SIZE*0.12,
-                       TF_SIZE*0.53,
-                       TF_SIZE*0.21,
-                       TF_SIZE*0.27,
-                       TF_SIZE*0.78};
+const int seekpos[] = {int(TF_SIZE*0.52),
+                       int(TF_SIZE*0.23),
+                       int(TF_SIZE*0.73),
+                       int(TF_SIZE*0.77),
+                       int(TF_SIZE*0.80),
+                       int(TF_SIZE*0.12),
+                       int(TF_SIZE*0.53),
+                       int(TF_SIZE*0.21),
+                       int(TF_SIZE*0.27),
+                       int(TF_SIZE*0.78)};
 
 const int sp_size = sizeof(seekpos)/sizeof(int);
